shared/requirements.txt
# Shared Python dependencies for all services
# Install with: pip install -r requirements.txt

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# Correct way - this installs psycopg with binary extras
psycopg[binary]>=3.1.0,<4.0

# Redis
redis==5.0.1

# Firebase

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# Configuration & Environment
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0

# Resilience & Retry Logic
tenacity==8.2.3

# HTTP Client
httpx==0.25.2

# Validation
email-validator==2.1.0

# Development dependencies (install with requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0
# mypy==1.7.1
# pre-commit==3.5.0
